BoardConfigCommon: stop using the legacy OMX hacks
What an awful mess this is. We don't need the legacy OMX stuff
(we have 4.1 BSP blobs), but we still have the legacy mmparser
(libmmparser.so).

I actually experimented with this a long time ago but it exploded
due to some unrelated error with NATIVE_WINDOW_SET_BUFFERS_SIZE.
Somehow legacy OMX avoided it and here we are.

Kudos to @Kra1o5 for trying it out despite me saying it doesn't
work. :P